+++ react/components/MinicartIconButton.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { FC } from 'react'
 import { ButtonWithIcon } from 'vtex.styleguide'
 import { IconCart } from 'vtex.store-icons'
 import { useOrderForm } from 'vtex.order-manager/OrderForm'
@@ -9,7 +9,13 @@ import styles from '../styles.css'
 
 const CSS_HANDLES = ['minicartIconContainer', 'minicartQuantityBadge'] as const
 
-const MinicartIconButton = () => {
+interface MinicartIconButtonProps {
+  quantityDisplay: MinicartIconButtonType
+}
+
+const MinicartIconButton: FC<MinicartIconButtonProps> = ({
+  quantityDisplay,
+}) => {
   const { orderForm, loading }: OrderFormContext = useOrderForm()
   const handles = useCssHandles(CSS_HANDLES)
   const { open, openBehavior, openOnHoverProp } = useMinicartState()
@@ -35,12 +41,16 @@ const MinicartIconButton = () => {
     dispatch({ type: open ? 'CLOSE_MINICART' : 'OPEN_MINICART' })
   }
 
+  const showQuantityBadge =
+    (itemQuantity > 0 && quantityDisplay === 'not-empty') ||
+    quantityDisplay === 'always'
+
   return (
     <ButtonWithIcon
       icon={
         <span className={`${handles.minicartIconContainer} gray relative`}>
           <IconCart />
-          {itemQuantity > 0 && (
+          {showQuantityBadge && (
             <span
               style={{ userSelect: 'none' }}
               className={`${handles.minicartQuantityBadge} ${styles.minicartQuantityBadgeDefault} c-on-emphasis absolute t-mini bg-emphasis br4 w1 h1 pa1 flex justify-center items-center lh-solid`}