+++ react/typings/global.ts
@@ -4,11 +4,15 @@ interface MinicartProps {
   linkVariationUrl: string
   maxDrawerWidth: number | string
   drawerSlideDirection: SlideDirectionType
+  quantityDisplay: MinicartIconButtonType
 }
+
+type MinicartIconButtonType = 'always' | 'never' | 'not-empty'
 
 type SlideDirectionType =
   | 'horizontal'
   | 'vertical'
   | 'rightToLeft'
   | 'leftToRight'
+
 type MinicartVariationType = 'popup' | 'drawer' | 'link'