+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Updated `CODEOWNERS` file with responsible teams for each directory.
 - Updated `.all-contributorsrc`.
+- Add `quantityDisplay` prop in `minicart.v2`.
 
 ## [2.43.5] - 2020-02-18
 ### Changed